// src/services/central-api/generated/model/crt-price-tier.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PL Cental API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CrtPriceTier
 */
export interface CrtPriceTier {
    /**
     * 
     * @type {number}
     * @memberof CrtPriceTier
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof CrtPriceTier
     */
    'name'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof CrtPriceTier
     */
    'minDay'?: number | null;
    /**
     * 
     * @type {number}
     * @memberof CrtPriceTier
     */
    'maxDay'?: number | null;
    /**
     * 
     * @type {number}
     * @memberof CrtPriceTier
     */
    'sequence'?: number | null;
    /**
     * 
     * @type {boolean}
     * @memberof CrtPriceTier
     */
    'isActive'?: boolean | null;
    /**
     * 
     * @type {string}
     * @memberof CrtPriceTier
     */
    'createBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CrtPriceTier
     */
    'createTimestamp'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CrtPriceTier
     */
    'updateBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CrtPriceTier
     */
    'updateTimestamp'?: string | null;
}