// src/services/central-api/generated/model/ctl-content.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PL Cental API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CtlContent
 */
export interface CtlContent {
    /**
     * 
     * @type {number}
     * @memberof CtlContent
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof CtlContent
     */
    'tenantCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlContent
     */
    'branchCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlContent
     */
    'metaKeyword'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlContent
     */
    'metaDescription'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlContent
     */
    'description'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlContent
     */
    'content'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlContent
     */
    'thumbmailUrl'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlContent
     */
    'slug'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlContent
     */
    'url'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlContent
     */
    'createBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlContent
     */
    'createTimestamp'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlContent
     */
    'updateBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlContent
     */
    'updateTimestamp'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlContent
     */
    'categoryCode'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof CtlContent
     */
    'isActive'?: boolean | null;
    /**
     * 
     * @type {boolean}
     * @memberof CtlContent
     */
    'isDelete'?: boolean | null;
    /**
     * 
     * @type {string}
     * @memberof CtlContent
     */
    'deleteBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlContent
     */
    'deleteTimestamp'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlContent
     */
    'status'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlContent
     */
    'publishDate'?: string | null;
}