// src/services/central-api/generated/model/vehicle-master-output.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PL Cental API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface VehicleMasterOutput
 */
export interface VehicleMasterOutput {
    /**
     * 
     * @type {number}
     * @memberof VehicleMasterOutput
     */
    'vehicleMasterId'?: number;
    /**
     * 
     * @type {string}
     * @memberof VehicleMasterOutput
     */
    'description'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof VehicleMasterOutput
     */
    'branchCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof VehicleMasterOutput
     */
    'brandName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof VehicleMasterOutput
     */
    'modelCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof VehicleMasterOutput
     */
    'modelName'?: string | null;
}